% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Dataset for NMAPropForest}
\format{
A data frame:
\describe{
  \item{study}{Study label}
  \item{id}{Study id}
  \item{t}{Treatment label}
  \item{r}{Number of events (for binary outcomes)}
  \item{n}{Sample size}
  \item{rob}{Risk of bias}
}
}
\source{
Example dataset included with the NMAPropForest package
}
\usage{
example_data
}
\description{
A small dataset used to demonstrate the usage of \code{NMAPropForest()}.
}
\keyword{datasets}
